be/requirements.txt
aioserial==1.3.1
annotated-types==0.7.0
anyio==4.7.0
click==8.1.8
fastapi==0.115.6
future==1.0.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
iso8601==2.1.0
lgpio==0.2.2.0
pydantic==2.10.4
pydantic_core==2.27.2
pyserial==3.5
PyYAML==6.0.2
rpi-lgpio==0.6
RPi.GPIO==0.7.1
RPLCD==1.3.1
smbus2==0.5.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
